# docs/swagger.yaml
definitions:
  models.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: List all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription info
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
      summary: Get a subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription info
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Get sum of subscriptions filtered by user/service/date
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Start YYYY-MM
        in: query
        name: start
        required: true
        type: string
      - description: End YYYY-MM
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      summary: Sum subscriptions
      tags:
      - subscriptions
swagger: "2.0"